+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `--convergencesets` for showing convergence sets definitions.
+- `--cell` and `--cell-registers` for showing cell registers definitions.
 
 ## [v0.5.1] - 2026-09-01